feat: prefered languages implementation (#6)
This pull request introduces changes to the user schema, the handling of
user language preferences, and database initialization scripts. The key
updates include adding validation for preferred languages and ensuring
consistency in language codes.

### User Schema and Validation Enhancements:
*
[`Backend/src/Schemas/userSchema.js`](diffhunk://#diff-e228add4ee5671b50e637f1830b023a5b401afa411e30f9ff579d2ee90fffa79R26-R27):
Added `acceptedLanguages` array to define valid language options and
updated the `userSchema` to include a `prefered_language` field with
validation and transformation rules.
[[1]](diffhunk://#diff-e228add4ee5671b50e637f1830b023a5b401afa411e30f9ff579d2ee90fffa79R26-R27)
[[2]](diffhunk://#diff-e228add4ee5671b50e637f1830b023a5b401afa411e30f9ff579d2ee90fffa79R140-R148)

### User Data Handling:
*
[`Backend/src/Utils/getPublicUser.js`](diffhunk://#diff-834f4cc9f91598910801757d4a850457dd4b496128b745c392b5ffbdc65e7791R13):
Updated the `getPublicUser` function to include the `prefered_language`
field in the returned user object.

### Database Initialization:
*
[`Docker/PostgreSQL/init/init.sql`](diffhunk://#diff-1cad65cb0ae9b2f586380e45ebf3748c081e43d844123bd5d091a971867a7a5fL3-R3):
Modified the `language_enum` type to use lowercase values and updated
the `prefered_language` field in the `users` table to use the new
lowercase enum values.
[[1]](diffhunk://#diff-1cad65cb0ae9b2f586380e45ebf3748c081e43d844123bd5d091a971867a7a5fL3-R3)
[[2]](diffhunk://#diff-1cad65cb0ae9b2f586380e45ebf3748c081e43d844123bd5d091a971867a7a5fL14-R14)

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>